# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CodeForge API
  description: |
    REST API for CodeForge -- AI coding agent orchestration platform.

    All endpoints under /api/v1 require a valid JWT bearer token unless
    explicitly marked as public (login, setup-status, health).

    Refresh tokens are delivered as httpOnly cookies on the /auth path.
  version: "0.8.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://localhost:8080
    description: Local development (root, for /health)
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List all projects
      operationId: listProjects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags: [Projects]
      summary: Create a new project
      description: |
        Creates a project. Requires admin or editor role.
        Optionally pass local_path to auto-adopt a local workspace.
        Config keys the backend reads: policy_preset, execution_mode (only
        "mount"), test_command and lint_command (quality gate commands; 400
        unless they split like a shell command line and start with an
        executable on the worker's allowlist).
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project by ID
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Projects]
      summary: Update a project
      description: Updates project metadata. Requires admin or editor role. Config keys are merged (null deletes one); test_command and lint_command are validated as on create (400).
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                repo_url:
                  type: string
                config:
                  type: object
                  description: |
                    Merge patch for the project config: a key with a string value is set, a key
                    with null is deleted, keys not in the request are kept. Missing, null or empty
                    config leaves the stored config unchanged.
                  additionalProperties:
                    type: string
                    nullable: true
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: Delete project (admin only)
      description: |
        With per-tenant tool users (`workspace.tool_acls: required`, production; KI-96) the project row is
        deleted together with a `workspace_deletions` record, and the worker removes the workspace afterwards
        as the tenant's tool UID (asynchronously, retried until done). The request is refused with 409 while
        the project has a run that has not ended, a conversation with an active turn, or a queued or running
        task. With `workspace.tool_acls: off` the Go Core removes the workspace itself.
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "204":
          description: Project deleted (the workspace is removed by the worker when tool ACLs are required)
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The project has active work (runs, conversation turns or tasks); stop it first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: The tenant has no tool UID yet and the tool UID range is exhausted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # -----------------------------------------------
  # Project files (workspace file API) and stack detection
  # -----------------------------------------------
...
  /detect-stack:
    post:
      tags: [Projects]
      summary: Detect the language stack of a directory
      description: |
        Scans a directory for languages, frameworks and tool recommendations. The path must be absolute and lie strictly
        inside the caller's own tenant area of the workspace root (`<workspace.root>/<tenant_id>/...`, KI-106); it is
        resolved inside that area, so no symlink leads out of it. Platform admins may also name a directory strictly
        inside a `workspace.adopt_roots` directory outside the workspace root. Any other path answers 400: another
        tenant's directory, the tenant directory itself, the workspace root, or a symlink that leaves the area.
      operationId: detectStackByPath
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Absolute directory path
      responses:
        "200":
          description: Detected stack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StackDetectionResult"
        "400":
          description: Path missing, not absolute, or outside the caller's tenant area (outside the workspace root, outside the adopt roots or the caller is no platform admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # -----------------------------------------------
  # Conversations
  # -----------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from POST /auth/login.
        Pass as: Authorization: Bearer <token>
  parameters:
    ProjectID:
      name: id
      in: path
      required: true
      description: Project UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Insufficient permissions (role check failed)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        repo_url:
          type: string
        status:
          type: string
          enum: [created, cloning, ready, error]
        config:
          type: object
          additionalProperties:
            type: string
          description: Project-level configuration (policy_preset, execution_mode, detected_languages, expansion_prompt, ...)
        workspace_path:
          type: string
          description: Absolute path to the local workspace (if adopted/cloned)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        repo_url:
          type: string
          description: Git clone URL (optional for local-only projects)
        local_path:
          type: string
          description: Local filesystem path to auto-adopt as workspace
        config:
          type: object
          additionalProperties:
            type: string
          description: |
            Project config map. Keys the backend reads: policy_preset, execution_mode (default exec
            mode for runs and agentic conversations; only mount is accepted until sandbox isolation
            exists), expansion_prompt. Autonomy comes from the selected mode.
...
    StackDetectionResult:
      type: object
      properties:
        languages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              confidence:
                type: number
              manifests:
                type: array
                items:
                  type: string
              frameworks:
                type: array
                items:
                  type: string
        recommendations:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              id:
                type: string
              name:
                type: string
              reason:
                type: string
        scanned_path:
          type: string
...